Fix package name in Cargo.toml and add initial tests for encoding functions
